+++ cmd/iconbridge/chain/algo/sender_test.go
@@ -3,114 +3,13 @@ package algo
 import (
 	"context"
 	"encoding/base64"
-	"encoding/json"
-	"fmt"
 	"testing"
 	"time"
 
-	"github.com/algorand/go-algorand-sdk/client/kmd"
-	"github.com/algorand/go-algorand-sdk/crypto"
 	"github.com/algorand/go-algorand-sdk/encoding/msgpack"
 	"github.com/bmizerany/assert"
 	"github.com/icon-project/icon-bridge/cmd/iconbridge/chain"
-	"github.com/icon-project/icon-bridge/common/log"
-	"github.com/icon-project/icon-bridge/common/wallet"
 )
-
-const sandboxAddress = "http://localhost:4001"
-const sandboxToken = "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-const (
-	KMD_ADDRESS         = "http://localhost:4002"
-	KMD_TOKEN           = "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-	KMD_WALLET_NAME     = "unencrypted-default-wallet"
-	KMD_WALLET_PASSWORD = ""
-)
-const approvalPath = "bmc/approval.teal"
-const clearPath = "bmc/clear.teal"
-
-func getAccounts() ([]crypto.Account, error) {
-	client, err := kmd.MakeClient(KMD_ADDRESS, KMD_TOKEN)
-	if err != nil {
-		return nil, fmt.Errorf("Failed to create client: %+v", err)
-	}
-
-	resp, err := client.ListWallets()
-	if err != nil {
-		return nil, fmt.Errorf("Failed to list wallets: %+v", err)
-	}
-
-	var walletId string
-	for _, wallet := range resp.Wallets {
-		if wallet.Name == KMD_WALLET_NAME {
-			walletId = wallet.ID
-		}
-	}
-
-	if walletId == "" {
-		return nil, fmt.Errorf("No wallet named %s", KMD_WALLET_NAME)
-	}
-
-	whResp, err := client.InitWalletHandle(walletId, KMD_WALLET_PASSWORD)
-	if err != nil {
-		return nil, fmt.Errorf("Failed to init wallet handle: %+v", err)
-	}
-
-	addrResp, err := client.ListKeys(whResp.WalletHandleToken)
-	if err != nil {
-		return nil, fmt.Errorf("Failed to list keys: %+v", err)
-	}
-
-	var accts []crypto.Account
-	for _, addr := range addrResp.Addresses {
-		expResp, err := client.ExportKey(whResp.WalletHandleToken, KMD_WALLET_PASSWORD, addr)
-		if err != nil {
-			return nil, fmt.Errorf("Failed to export key: %+v", err)
-		}
-
-		acct, err := crypto.AccountFromPrivateKey(expResp.PrivateKey)
-		if err != nil {
-			return nil, fmt.Errorf("Failed to create account from private key: %+v", err)
-		}
-
-		accts = append(accts, acct)
-	}
-
-	return accts, nil
-}
-
-func createTestSender() (chain.Sender, error) {
-	accts, err := getAccounts()
-	if err != nil {
-		return nil, fmt.Errorf("Error generating KMD account: %v", err)
-	}
-	account := accts[0]
-
-	algodAccess := []string{sandboxAddress, sandboxToken}
-	ctx, _ := context.WithTimeout(context.Background(), 60*time.Second)
-
-	appId, err := deployContract(ctx, algodAccess, [2]string{approvalPath, clearPath}, account)
-	if err != nil {
-		return nil, fmt.Errorf("Error deploying BMC: %v", err)
-	}
-	opts := map[string]interface{}{"app_id": appId}
-	rawOpts, err := json.Marshal(opts)
-	if err != nil {
-		return nil, fmt.Errorf("Marshalling opts: %v", err)
-	}
-	w, err := wallet.NewAvmWalletFromPrivateKey(&account.PrivateKey)
-	if err != nil {
-		return nil, fmt.Errorf("Couldn't create wallet: %v", err)
-	}
-
-	s, err := NewSender(
-		chain.BTPAddress(icon_bmc), chain.BTPAddress(algo_bmc),
-		algodAccess, w,
-		rawOpts, log.New())
-	if err != nil {
-		return nil, fmt.Errorf("Error creating new sender: %v", err)
-	}
-	return s, nil
-}
 
 func Test_Abi(t *testing.T) {
 	s, err := createTestSender()
@@ -127,7 +26,6 @@ func Test_Abi(t *testing.T) {
 		t.Logf("Failed calling abi:%v", err)
 		t.FailNow()
 	}
-	fmt.Println(ret)
 	concatString := ret.MethodResults[0].ReturnValue.(string)
 	assert.Equal(t, concatString, "thisstringisjoined")
 }
@@ -167,15 +65,12 @@ func Test_Segment(t *testing.T) {
 				},
 			}},
 	}
-	tx, newmsg, err := s.Segment(ctx, msg)
+	tx, _, err := s.Segment(ctx, msg)
 
 	if err != nil {
 		t.Logf("Couldn't segment message:%v", err)
 		t.FailNow()
 	}
-	fmt.Println(tx)
-	fmt.Println("......................")
-	fmt.Println(newmsg)
 
 	sss := tx.(*relayTx).msg
 
@@ -194,6 +89,8 @@ func Test_Segment(t *testing.T) {
 
 		recSli = append(recSli, finalRcp)
 	}
-	fmt.Println("......................")
 
+	for i := range recSli {
+		assert.Equal(t, *msg.Receipts[i], recSli[i])
+	}
 }
